ocr_server/car_certification/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.112.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
joblib==1.4.2
numpy==2.0.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
setuptools==72.1.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
wheel==0.43.0
